requirements.txt
Django>=4.2
psycopg2-binary>=2.9
django-crispy-forms
crispy-bootstrap5
